boya/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy
scikit-learn
matplotlib
seaborn
pillow
PyYAML
tensorboard
